requirements.txt
Django==2.2.10
django-allauth==0.42.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
#djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.5
python3-openid==3.2.0
requests==2.24.0
